# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "使用go语言搭建的管理后端",
        "title": "四零专属管理后台",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/system/rolePermissionRel/assign": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rolePermissionRelController"
                ],
                "summary": "为角色分配权限",
                "operationId": "assignPermissionsToRole",
                "parameters": [
                    {
                        "description": "分配权限参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.RolePermissionRelAssignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
        "/system/rolePermissionRel/listByPermission": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rolePermissionRelController"
                ],
                "summary": "根据权限ID获取角色列表",
                "operationId": "getRolesByPermissionId",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "权限ID",
                        "name": "permissionId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-array_system_RoleWithPermissionVo"
                        }
                    }
                }
            }
        },
        "/system/rolePermissionRel/listByRole": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rolePermissionRelController"
                ],
                "summary": "根据角色ID获取权限列表",
                "operationId": "getPermissionsByRoleId",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "roleId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-array_system_RolePermissionRelVo"
                        }
                    }
                }
            }
        },
        "/system/rolePermissionRel/remove": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rolePermissionRelController"
                ],
                "summary": "移除角色的权限",
                "operationId": "removePermissionFromRole",
                "parameters": [
                    {
                        "description": "移除权限参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.RolePermissionRelRemoveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
        "/system/user/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userController"
                ],
                "summary": "创建用户",
                "operationId": "createUser",
                "parameters": [
                    {
                        "description": "创建用户参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.UserCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
        "/system/user/delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userController"
                ],
                "summary": "删除用户",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "description": "删除用户参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.UserDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
        "/system/user/detail": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userController"
                ],
                "summary": "获取用户详情",
                "operationId": "getUserDetail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-system_UserVo"
                        }
                    }
                }
            }
        },
        "/system/user/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userController"
                ],
                "summary": "获取用户列表",
                "operationId": "getUserList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账号",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "昵称",
                        "name": "nickName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-response_PageResult-system_UserVo"
                        }
                    }
                }
            }
        },
        "/system/user/update": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userController"
                ],
                "summary": "更新用户",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "description": "更新用户参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.UserUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
        "/system/user/updatePwd": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userController"
                ],
                "summary": "修改密码",
                "operationId": "updatePassword",
                "parameters": [
                    {
                        "description": "修改密码参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.UserUpdatePwdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResult-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.ApiResult-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
        "response.ApiResult-array_system_RolePermissionRelVo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.RolePermissionRelVo"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
        "response.ApiResult-array_system_RoleWithPermissionVo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.RoleWithPermissionVo"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
...
        "response.ApiResult-response_PageResult-system_UserVo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/response.PageResult-system_UserVo"
                },
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
...
        "response.ApiResult-system_UserVo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/system.UserVo"
                },
                "msg": {
                    "type": "string",
                    "example": "成功"
                }
            }
        },
...
        "response.PageResult-system_UserVo": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.UserVo"
                    }
                },
                "pageNum": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "system.RolePermissionRelAssignRequest": {
            "type": "object",
            "required": [
                "permissionIds",
                "roleId"
            ],
            "properties": {
                "permissionIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "roleId": {
                    "type": "integer"
                }
            }
        },
        "system.RolePermissionRelRemoveRequest": {
            "type": "object",
            "required": [
                "permissionId",
                "roleId"
            ],
            "properties": {
                "permissionId": {
                    "type": "integer"
                },
                "roleId": {
                    "type": "integer"
                }
            }
        },
        "system.RolePermissionRelVo": {
            "type": "object",
            "properties": {
                "permissionId": {
                    "type": "integer"
                },
                "permissionName": {
                    "type": "string"
                },
                "perms": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "system.RoleWithPermissionVo": {
            "type": "object",
            "properties": {
                "roleId": {
                    "type": "integer"
                },
                "roleKey": {
                    "type": "string"
                },
                "roleName": {
                    "type": "string"
                }
            }
        },
        "system.UserCreateRequest": {
            "type": "object",
            "required": [
                "account",
                "nickName",
                "password"
            ],
            "properties": {
                "account": {
                    "type": "string"
                },
                "avatarId": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "system.UserDeleteRequest": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "system.UserUpdatePwdRequest": {
            "type": "object",
            "required": [
                "newPassword",
                "password",
                "userId"
            ],
            "properties": {
                "newPassword": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "system.UserUpdateRequest": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "avatarId": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "system.UserVo": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "avatarId": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}